broadwell: Fix TCO register size and event reporting
TCO registers are 16bit not 32bit.  Also do not log the
TCO reset event in S3 resume path to avoid it being logged
when TCO is not actually tripping.

BUG=chrome-os-partner:28234
BRANCH=None
TEST=manual:
1) build and boot on samus
2) modify kernel command line with nmi_watchdog=0
3) while sleep 1 ; do echo -n V ; done > /dev/watchdog &
4) fg 1
5) ctrl-Z
6) wait for reboot
7) check event log for TCO event
8) check suspend/resume path to ensure no TCO event logged

Original-Change-Id: I9cd8627de8498b280deb088f3a8e1e20546e2f96
Original-Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/211840
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
(cherry picked from commit 5952fe4672d07bd39e345f2048c2bfc510bf9f2a)

Reviewed-on: http://review.coreboot.org/8954
Tested-by: build bot (Jenkins)

+++ src/soc/intel/broadwell/broadwell/pm.h
@@ -123,8 +123,8 @@ struct chipset_power_state {
 	uint16_t pm1_sts;
 	uint16_t pm1_en;
 	uint32_t pm1_cnt;
-	uint32_t tco1_sts;
-	uint32_t tco2_sts;
+	uint16_t tco1_sts;
+	uint16_t tco2_sts;
 	uint32_t gpe0_sts[4];
 	uint32_t gpe0_en[4];
 	uint16_t gen_pmcon1;
